+++ AdminApi/DTO/App/TempleDTO/CreateTempleDTO.cs
@@ -18,7 +18,7 @@ namespace AdminApi.DTO.App.TempleDTO
 
     public class UpdateTempleDTO
     {
-        public Int64 TempleId { get; set; }
+        public int TempleId { get; set; }
         public string TempleName { get; set; }
         public int CountryId { get; set; }
         public int StateId { get; set; }
